pyproject.toml
[tool.poetry]
name = "doc-flash"
version = "0.1.0"
description = "FastAPI service for code documentation with pluggable LLM providers (Gemini-first)."
authors = ["DocFlash Team <dev@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "doc_flash", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
google-generativeai = "^0.7.0"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
